% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp.R
\name{harp}
\alias{harp}
\title{Perform allele-specific alignment with fastq files}
\usage{
harp(fastqFiles1, fastqFiles2 = NULL, index1prefix, index2prefix,
  minAS = -20, minQual = 20, threads = getOption("threads", 1L),
  fields = c("AS", "XM", "XO", "XG", "NM"), ...)
}
\arguments{
\item{fastqFiles1}{A vector of relative/absolute paths to fastq files. Each fastq file provided is parsed independently.}

\item{fastqFiles2}{A vector of read2 fastq files corresponding to the mate-pairs of the files defined in fastqFiles1.}

\item{index1prefix}{A relative/absolute path to a bowtie2 index of one of the reference haplotypes. Note that the prefix of reference 1 and 2 should be different (e.g., not both be "genome")}

\item{index2prefix}{A relative/absolute path to a bowtie2 index of the second reference haplotype.}

\item{minAS}{An integer specifying the minimum alignment score to consider a read as aligned.}

\item{minQual}{A positive integer specifying the nimimum alignment score to parse a read.}

\item{threads}{A positive integer specifying the number of threads to use for alignment and parsing.}

\item{fields}{A vector of optional sam fields to preserve in the output sam files.}
}
\value{
A vector of file names of the parsed sam files.
}
\description{
Perform allele-specific alignment with fastq files
}
