% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp_pipeline.R
\name{harp_pipeline}
\alias{harp_pipeline}
\title{harp_pipeline}
\usage{
harp_pipeline(
  ...,
  train_data,
  cell_reference_profile,
  bulk_data = NULL,
  n_folds = 5,
  lambda_seq = c(seq(0, 1, by = 0.1), 2^seq(1, 5, by = 1)),
  verbose = TRUE
)
}
\arguments{
\item{train_data}{a list containing training bulk expression data with "mixtures" as expression values and the corresponding cell counts as "quantities."}

\item{cell_reference_profile}{a matrix containing the cell profiles, with genes in rows and unique cell types in columns.}

\item{bulk_data}{a matrix containing the bulk expression data to be deconvolved, with genes in rows and samples in columns.}

\item{n_folds}{integer specifying the number of buckets/folds in the cross-validation. for example, if you have 15 samples in your training data and want to
perform 5-fold cross-validation, each bucket will have 3 samples.}

\item{lambda_seq}{a numeric value (if no cross-validation is performed) or a sequence of values for regularization.}
}
\value{
return.list list
}
\description{
this script is to train harp to infer all parameteres (lambda, estimated refrence, alpha, g in dtd) and then deconvolve the bulk data
}
