% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_config.R
\name{make_point_verif_config}
\alias{make_point_verif_config}
\alias{edit_config}
\title{Verification configuration files}
\usage{
make_point_verif_config(config_file_name, edit = TRUE)

edit_config(config_file_name)
}
\arguments{
\item{config_file_name}{The name of the configuration file to be made}

\item{edit}{Logical. Whether to open the file in an editor.}
}
\description{
These functions help you to make, edit and read configuration files for
verification in harp. `make_*_verif_config()` will open an example
configuration that you can edit and save. `edit_config()` opens an editor
for an existing file. More documentation about the variables in the
configuration file can found in \code{\link{run_point_verif()}}.
}
\details{
In _RStudio_ the file will be opened in the _Source_ pain, otherwise your
default editor will be used.
}
\seealso{
[run_point_verif()]
}
