% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_verif_param.R
\name{make_scaling}
\alias{make_scaling}
\title{Make a list for scaling forecasts or observations}
\usage{
make_scaling(scaling, new_units, mult = FALSE)
}
\arguments{
\item{scaling}{The scaling to apply to the data. By default the scaling is
additive, but if \code{mult = TRUE} it is multiplicative.}

\item{new_units}{The name of the new units. If missing, the units name will
be unchanged.}

\item{mult}{Logical. Whether the scaling is multiplicative. The default is
\code{FALSE}, meaning that the scaling is additive.}
}
\description{
`make_scaling()` is a helper function to be used when defining the scaling
for verification parameters with \code{\link{make_verif_param}}.
}
