% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_verif_params.R
\name{make_verif_params_file}
\alias{make_verif_params_file}
\title{Copy and open an example verification parameters file}
\usage{
make_verif_params_file(file_name, type = c("R", "json", "txt"))
}
\arguments{
\item{file_name}{The name of the file you want to make. It should either
have no extension, or an extension that matches `type`.}

\item{type}{The type of file to create. Can be "R" (the default), "json", or
"txt".}
}
\description{
Create a new verification parameters from an example. You can choose whether
to have this file as an R script (the recommended type), a json file or a
text file. The example files are quite comprehensive so should be edited
to your needs.
}
