% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_config.R
\name{new_point_verif_project}
\alias{new_point_verif_project}
\title{Create a new point verification project}
\usage{
new_point_verif_project(
  dir,
  permission = "755",
  config_file = "harp_point_verif_conf.txt",
  params_file = "harp_point_verif_params.R",
  script_file = "harp_point_verify.R"
)
}
\arguments{
\item{dir}{The directory for the verification project. If it does not exist,
you will prompted to confirm that you want to create the directory.}

\item{permission}{The permissions for `dir`. The default is 755.}

\item{config_file}{The name of the configuration file. This defaults to
`harp_point_verif_config.txt`.}

\item{params_file}{The name of the parameter definitions file. This defaults
to `harp_point_verif_params.R`, which contains R code using helper
functions to define some aspects of the parameters. You may also give a
file name with a `.json` or a `.txt` extension, in which case the params
file will be in that format.}

\item{script_file}{The name of the script that will run the verification.
This defaults to `harp_point_verify.R` and in most cases should not be
edited once it is created.}
}
\description{
This function copies all the files that you need to run a point verification
job from a shell terminal into the specified directory. This includes a
parameter definitions file, a configuration file, the run script and a
README. The names of these files can optionally be changed, though it is
recommended to use the defaults until you become accustomed with the system.
You can also optionally automatically open the necessary files for editing.
}
