% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_config.R
\name{read_config}
\alias{read_config}
\title{Read a configuration file}
\usage{
read_config(config_file_name)
}
\arguments{
\item{config_file_name}{The full path to the configuration file}
}
\value{
A named list
}
\description{
A configuration file is used to provide information for a verification
run, such as paths, file templates, file formats, and other basic settings.
}
