% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_verif_params.R
\name{read_verif_params}
\alias{read_verif_params}
\title{Verification parameter files}
\usage{
read_verif_params(file_name)
}
\arguments{
\item{file_name}{The name of the parameters file}
}
\value{
A named list
}
\description{
`read_verif_params()` reads a parameter file for verification. This can be a
file that contains R code, a text file or a json file.
}
\examples{
read_verif_params(system.file("config", "define_params.R"))
read_verif_params(system.file("config", "params.txt"))
read_verif_params(system.file("config", "params.json"))
}
