% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_verif_param.R
\name{write_verif_params}
\alias{write_verif_params}
\title{Write parameter definitions to a file}
\usage{
write_verif_params(param, file_name, file_type = c("R", "rds", "json"))
}
\arguments{
\item{param}{A list containing one or more parameter created using
\code{\link{make_verif_param}()}}

\item{file_name}{The full path to the file where the parameters are to be
saved.}

\item{file_type}{The type of file to save. Can be "R" for an R source file,
"rds", or "json". If one of these is the file extension in `file_name`
then `file_type` is ignored.}
}
\description{
Once parameter definitions have been made using
\code{\link{make_verif_param}()} `write_verif_params()` can be used to
save them to a file for future use. You have a choice to save as R source
code, or JSON such that the files are editable outside of R, or as an rds
file that can be read into an R session.
}
