% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_verif_param.R
\name{make_verif_defaults}
\alias{make_verif_defaults}
\title{Make defualts for point verification}
\usage{
make_verif_defaults(verif_groups = "lead_time", obs_error_sd = NULL)
}
\arguments{
\item{verif_groups}{The groups for which to compute verification scores. That
is to say the stratification of the verification scores.
\code{\link[harpCore]{make_verif_groups}()} can be used to define the
verification groups. The default is for the verification to be grouped by
lead time.}

\item{obs_error_sd}{The maximum number of standard deviations that an
observation can be different from a forecast. Used as a sort of
represntativeness check. See
\code{\link[harpPoint]{check_obs_against_fcst}()} for more details.}
}
\description{
`make_verif_defaults()` can be used to set default values for point
verification regardless of the parameter. This is currently restricted to
verification groups and the number of standard deviations allowed between
observations and forecasts for observations to be accepted into the
verification. All other options are set at the parameter level using
\code{\link{make_verif_param()}}
}
